+++ docs/changelog.md
@@ -19,6 +19,8 @@
 
  - [bugfix] Fix "too many values to unpack" exception when scanning video
  - [bugfix] Fix incorrect calculation of total frames when showing progress
+ - [general] Show alert on any errors during motion scanning
+
 
 ----------------------------------------------------------